bin/magic-cli: Add binary and test that it runs
Note that we use the `#!/bin/sh` shebang. On some systems (such as
BSDs and Debian) this runs a non-Bash Bourne-like shell; on others
(CentOS?) it runs Bash.

We should try to deal with this in testing at some point later,
ensuring we work under both "generic" Bourne shell and Bash. This
might be a bit tricky given that our test infrastructure (BATS) runs
only under Bash, but we'll work out something.

If you're following along in the commit log as a learning experience,
keep the tests of this commit but modify it otherwise as the following
and see what the test results show you:
  1. When `bin/magic-cli` doesn't exist
  2. When it exists but isn't executable
  3. When it exists and runs but exits with a non-zero value
  3. When it exits with a zero value but doesn't print "Write me!"