Add the very basics of running a sub-command
We export some obvious environment variables, though we've not yet
decided if these are offical API level or not.

There are also a few things (mostly related to the full path to the
command dir) that we don't check.

XXX We really should be finding subcommands anywhere in the path, not
just in the same dir as the subcommand runner.