Rename from "sh-magic-cli" to "shubcommander"
So, I was trying to think of a better name for this sort of thing,
"subcommander" (kinda obvious, I know) came to mind, and so I did a
quick web and GitHub search (https://github.com/search?q=subcommander)
to see if it might be confused with any other project.

As it turns out, it will be confused with about a million of them. It
seems that the Slackers didn't bother to mention that this Wonderful
New Idea[1] of theirs has been done many times before (or maybe they
just couldn't be bothered to look around to see that).

[1]: https://slack.engineering/the-joy-of-internal-tools-4a1bb5fe905b

Anyway, I'll carry on with this for a bit, but my new name will be
"shubcommander" for several reasons:

  1. It differentiates it from all the "subcommander" projects out
     there.
  2. It's written in `sh`, duh.
  3. I have a weird predilection for a post-consonant 'h' as the
     second letter of a project name.

Now, why all this explanation? Normally I'd just rewrite the branch
from the start and make it look like I'd known all this from the
start. However, I'm Building Git Repos for Didactic Purposes these
days (put it down to not enough chance to do pair programming), and I
think this can be a valuable lesson.