Detect no parms and start to work on help message
We currently just print a generic usage message with a placeholder for
the subcommand names. The next step is to print the list of
subcommands, and probably a bit of usage help for them, too.

Just as a side note, having automated tests makes moving forward a lot
easier; they've already caught two or three things I'd broken in other
areas when trying to implement something as simple as this commit.